Enhance data source component (#583)
