+++ src/components/databaseDocs/styles.css
@@ -7,12 +7,30 @@
   padding: 12px;
   border-radius: 12px;
   box-shadow: var(--ifm-global-shadow-md);
+  transition: all 0.3s ease;
+}
+
+.minified {
+  position: sticky;
+  top: 120px;
+  right: 0;
+  padding: 8px;
+  height: 140px;
+  z-index: 1000;
+  background-color: var(--main-bg-color);
+  box-shadow: 0 -40px 0 5px var(--main-bg-color);
 }
 
 .button-wrapper {
-  display: flex;
-  flex-wrap: wrap;
+  display: inline-flex;
   gap: 15px;
+  align-items: center;
+  overflow-x: auto;
+  white-space: nowrap;
+}
+
+.picker-wrapper.minified .button-wrapper {
+  gap: 8px;
 }
 
 .data-source {
@@ -24,7 +42,28 @@
   cursor: pointer;
   width: auto;
   align-items: center;
+  text-align: center;
   text-decoration: none !important;
+}
+
+.picker-wrapper.minified .data-source {
+  margin-top: 2px;
+  margin-bottom: 2px;
+}
+
+.picker-wrapper.minified .data-source.selected {
+  border-width: 1.6px;
+}
+
+.picker-wrapper.minified .image-container {
+  height: 40px;
+  width: 40px;
+}
+
+.picker-wrapper.minified .image-container img,
+.picker-wrapper.minified .image-container svg {
+  width: 100%;
+  height: 100%;
 }
 
 .data-source p {
@@ -44,7 +83,8 @@
 }
 
 .data-source img {
-  max-width: 50px;
+  max-width: 100%;
+  max-height: 100%;
   height: auto;
   border-radius: 0;
 }
