+++ src/components/databaseDocs/utils.ts
@@ -1,0 +1,12 @@
+export const isBusinessLogicConnector = (savedPreference: string) =>
+  ['TypeScript', 'Python', 'Go'].includes(savedPreference);
+
+export const savePreference = (preference: string, history: any) => {
+  localStorage.setItem('hasuraV3ConnectorPreference', preference);
+
+  history.push({
+    search: `db=${preference}`,
+  });
+
+  return preference;
+};
